requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bson==0.5.10
DateTime==5.2
dnspython==2.4.2
fastapi==0.103.2
idna==3.4
motor==3.3.1
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing==3.7.4.3
typing_extensions==4.8.0
zope.interface==6.1
